- Limpado endpoints e serviços que o front não usa mais (pagamentos CRUD, track CRUD, consultas de usuário por CPF, listagem de reservas por pista). Atualizei o OpenAPI e removi os trechos correspondentes em repositórios/serviços/tests - Adicionado novos testes em SchedulingServiceTest para cobrir findSchedulingById, listSchedulings, listSchedulingsByUserId e checkAvailability - Ajustado o tipo de Track.id para Long em toda a aplicação (entidades, DTOs, projections, tests) para eliminar usos de Integer - Alterado o limite de 25 reservas por turno/dia removendo a dependência do track na consulta (checkAvailability agora usa countByScheduledDateAndTurno)

+++ src/main/java/com/projeto/demo/dto/SchedulingTrackSummaryDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class SchedulingTrackSummaryDto {
 
-    private Integer trackId;
+    private Long trackId;
     private String trackName;
     private Long total;
 }
